Upload the first time.
In the book "Introduction to Robotics, Mechanics and Control, the 3rd" by J.J.Craig, on Page64, we've got a series of formulas by calculating the function of a series of DH (Denavit and Hartenberg) parameters including the ALPHA angle, THETA angle, D and A. The input could be a DH kinematic table, matlab will generate its fomula using simple codes. And we could use subs function to calculate the specific number of each parameter. The final aim is to get the Transformation Matrix, which is commonly used to transform the coordinate in the last coordinate system to the base coordinate system. This is a very beautiful math tool.